+++ packages/ap2/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @docknetwork/ap2
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/crypto-utils@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes